+++ uniffi_bindgen/src/bindings/kotlin/test.rs
@@ -3,8 +3,9 @@ License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 use crate::{
-    bail, bindings::RunScriptOptions, cargo_metadata::CrateConfigSupplier,
-    library_mode::generate_bindings, Context, Result,
+    bail,
+    bindings::{generate, GenerateOptions, RunScriptOptions, TargetLanguage},
+    Context, Result,
 };
 use camino::{Utf8Path, Utf8PathBuf};
 use std::env;
@@ -37,15 +38,13 @@ pub fn run_script(
     let out_dir = test_helper.create_out_dir(tmp_dir, script_path)?;
     let cdylib_path = test_helper.copy_cdylib_to_out_dir(&out_dir)?;
 
-    generate_bindings(
-        &cdylib_path,
-        None,
-        &super::KotlinBindingGenerator,
-        &CrateConfigSupplier::from(test_helper.cargo_metadata()),
-        None,
-        &out_dir,
-        false,
-    )?;
+    generate(GenerateOptions {
+        languages: vec![TargetLanguage::Kotlin],
+        source: cdylib_path.to_path_buf(),
+        out_dir: out_dir.to_path_buf(),
+        crate_filter: Some(crate_name.to_string()),
+        ..GenerateOptions::default()
+    })?;
     let jar_file = build_jar(crate_name, &out_dir, options)?;
 
     let mut command = kotlinc_command(options);